moxui: add docs favicon

+++ scripts/tools/releaseDocs.ts
@@ -5,9 +5,20 @@ shelljs.exec(`pnpm docs:build`);
 
 // 上传
 shelljs.cd(`dist/docs`);
-shelljs.exec(`git init`);
-shelljs.rm("-rf", ".git");
-shelljs.exec(`git remote add origin git@mooui.com:mooui/mooui.github.io.git`);
-shelljs.exec(`git add .`);
-shelljs.exec(`git commit -m "moxui docs: release ${new Date().toISOString()}"`);
-shelljs.exec(`git push -u origin master --force`);
+
+// 有.git
+if (shelljs.test("-e", ".git")) {
+  shelljs.exec(`git add .`);
+  shelljs.exec(
+    `git commit -m "moxui docs: release ${new Date().toISOString()}"`
+  );
+  shelljs.exec(`git push`);
+} else {
+  shelljs.exec(`git init`);
+  shelljs.exec(`git remote add origin git@mooui.com:mooui/mooui.github.io.git`);
+  shelljs.exec(`git add .`);
+  shelljs.exec(
+    `git commit -m "moxui docs: release ${new Date().toISOString()}"`
+  );
+  shelljs.exec(`git push -u origin master --force`);
+}
